descompactacao muito doida, verificar os indices

+++ arvore.h
@@ -22,6 +22,6 @@ void desalocaCelula(void *c);
 bitmap *salvaArvore(tArvore *r, bitmap *bm);
 unsigned char getBitString(unsigned char *string,long index);
 tArvore *leArvore(unsigned char *string,tArvore *r,long *bitslidos);
-void imprime(unsigned char *string, long tamanho,FILE *f,tArvore *r,int *lidos);
+void imprime(unsigned char *string, long tamanho,FILE *f,tArvore *r,long *lidos);
 
 #endif
